test: TLS cipher list, create directories prior to docker
The `docker` process itself is being run as root, so volume mounts are still creating directories with root ownership. This can be avoided when the directories already exist.

Added `cd` commands for supporting `testssl.sh` outside of docker, if it's switched to that. `|| exit` is from `shellcheck` linting suggestion when using `cd`.

`rm -rf` had it's shared variable path swapped for hard-coded path, just as an added precaution due to risk that command can impose on a user running tests locally.