Collapse dh method `ONE_DIR=1` conditional branch
This only needs to modify the `DH_CUSTOM` path var. More DRY, the conditional branch for `ONE_DIR=0` (else), is kept and shifts it's indentation one level to the left.

I don't see any value in checking the alternative service for dh params file to copy over, so that's now dropped too.