fix: Spam bounced test copy/paste typo
Looks like when the 2nd (B) test case was added, they originally shared the same PRIVATE_CONFIG mount (before PRIVATE_CONFIG was a thing), and a copy/paste error was made with the undefined container name.

Fixed the container name to be distinct, and made the two PRIVATE_CONFIG distinct (A and B).