`ssl.md` - Better document `SSL_TYPE`
Reference `environment.md` early on the page as it better details the different types in a more terse overview.

Likewise, have the `environment.md` entry better link to relevant `SSL_TYPE` config. Clarify that `custom` isn't officially supported and further discourage `self-signed` certificates while encouraging `letsencrypt` and `manual` as preferable choices.

Brief addition to `ssl.md` about the dual cert support which was only touched on in `environment.md` (with insufficient indentation to render as a sub-list to `manual`.