fix: SSL_DOMAIN should not silently skip
Without quote wrapping the variables in the for declaration, `*.example.test` did not seem to be attempted. `mail.example.test` would be tried followed by `example.test` if no match was found.

A little surprised that shellcheck lint did not pick up on that. I have opted to use an array instead where the lint will advise quote wrapping vars.

This bug was assumed earlier due to a wildcard test where I would check the log for a failure about `mail.example.test` which should not actually pop up.

I could write another variant of the method to test that the failure does not get output instead, but have chosen not to as removing `SSL_DOMAIN` support in future is likely.