fix: Fail when build aborts from broken links
Using `set -ex` will exit the script as soon as a non-zero exit code is returned, such as when the docker image fails building the docs due to `build --strict` catching broken links.

This removes the need for `|| exit`. `-x` is used to improve command visibility in the workflow log output.

`--name` isn't necessary, but was part of the original PR. I've chosen not to include `-o pipefail` only because no pipes are used in the present script.