feat: Python extractor as utility script
The extraction python code is effectively the same, except for what is now `args.requested` var.

Extracted out into a python script that can be treated as a utility in the `$PATH` like other helper scripts. It can now be used and optionally tested directly instead of via `helper-functions.sh`.

Added an arg parser that provides better experience with help output and error/parsing checks out of the box. Should be easier to maintain now.

Updated to be compatible with Python 3, required for upgrade to Debian 11 Bullseye.