fix: SSL_DOMAIN should not silently skip
Without quote wrapping the variables in the for declaration, `*.example.test` did not seem to be attempted. `mail.example.test` would be tried followed by `example.test` if no match was found.

I'm a little surprised that shellcheck lint did not pick up on that. I have opted to use an array instead where the lint will actually advise to quote wrap vars.