chore: Re-apply extraction and sync of `_setup_dovecot_local_user`
PR merged, used merge commit to align with master and now it's good to restore changes for this method, along with syncing the changes from the PR it was waiting on.

Most of that PR has been assigned to a function to scope it, rather than blend in with the main account function logic.