Increased sleep duration for check-for-changes.sh
The script does not need to run every single second. This is not to say
that there is huge resource usage due to running every second, but every
ten seconds is more than enough. This _should_ act like a plaster when
it comes to #2348 - I KNOW THIS IS **NOT** A FIX. But this way, people
can first of all stay on Debian 11 and DMS 10.4.0, secondly, we may get
some more feedback on this mysterious issue due to people staying in DMS
10.4.0.