streamlined all cron additions
We now use HEREDOCs to add new files to cron. This is, after testing the
alternatives of copying files, the best solution in my opinion. We have
munite control over the contents (especially when the content of the
cron files depends on environment variables). Moreover, we need not
adjust the `Dockerfile`.