reworked log completely
This is a bigger commit - it reworks the whole log functionality. The
old one was confusing and not adhering to common standards, i.e. not
following convention over configuration. With this commit, we use the
common standard log levels

4. TRACE
3. DEBUG
2. INFO
1. WARN
0. ERROR

to better provide the information to users and maintainers. With this
changes, `DMS_DEBUG` was removed and `LOG_LEVEL` introduced. `DMS_DEBUG`
is not usefule anymore (obviously), and replaced by the new `LOG_LEVEL`,
which defaults to `info`.

All scripts and tests were adjusted to use the new format. There is a
special log level (`always`) which is used only by `start-mailserver.sh`
to show the welcome messages independent of the log level.

I took some care when converting the "old" log levels. With this rework,
it should be easier to identify problems in logs and provide users with
a clear and straight forward log level variable which is easy to
understand.

This is not actually a breaking change, but I thought it is fitting for
the next major release `v11`.