updated format for log level `always`
The new format is in my opinion more fitting for the two messages that
are logged by `start-mailserver.sh` (looks nice and clear to me).

Because Supervisor still has an open issue on color (8 years old...),
there is actually not much point in color codes, bold, etc., but this
way we're future-proofing ourself and we may someday see color.