removed log level `always`
After giving it a second thought, I figured it was useless to needlessly
complicate the log function with an additional log level when we can
just use `echo` for the two special instances. This also removes the
incentive to (ab)use this log level. All messages, except the two in
`start-mailserver.sh`, should have a proper log level.