corrected unclear circumstance with env variable in script
The env variable `POSTMASTER_ADDRESS` was set in a way that is not very
clear to the reader and indeed possibly wrong. The old way used the
default while the new way uses the correct variable from
`/root/.bashrc`.

Therefore, a new function `_get_dms_environment_variables` was
introduced which handles exporting vaiables from `/root/.bashrc`. This
could be used in the future as well.

I KNOW THIS IS NOT ACTUALLY RELATED TO THE LOG - PLEASE DON'T KILL ME :D
I just had to fix this... :) This is a small addition, I hope you don't
mind.