more efficient wait
`check-for-changes.sh` should now wait more efficiently on startup.
Moreover, the `mail_ssl_letsencrypt.bats` test was adjusted to view more
log. This has become necessary as the log grew in absolute size with
this PR (the messages are longer, more bytes to log).

I CAN'T FOR THE LIFE OF ME find the reason the test in
`mail_ssl_letsencrypt.bats` in the function `_should_extract_on_changes`
which is called by `_acme_wildcard` in the test named
`ssl(letsencrypt): Traefik 'acme.json' (*.example.test)` in the file
`mail_ssl_letsencrypt.bats` fails. @polarthene I need your help on this:
the last `cp` of the wildcard ACME does not seem to cause the message
`'/etc/letsencrypt/acme.json' has changed - extracting certifcates now`
to be logged, though it should, right? Maybe this was overlooked in the
past (false positive by viewing to much of the previous log???). In all
of my tests, the wildcard ACME does not cause the log message about the
file to be logged again...