adjust tests accordingly
Let'sEncrypt and SSL manual tests were adjusted to the new log.
Moreover,

```
assert_output --partial 'Change detected'
```

was dropped because it appears first in the logs, and due to the now
more verbose log, we'd need to increase the lenght of the `tail`ed
output, but this an unreliable method. I therefore simply dropped this
message, and I think this is fine, because the other messages are still
fine for testing whether all is well with the script and its
functionality.