adjust the script and documentation
The new way of providing your custom Postfix configuration is similar to
the old way, but now, the whole file is appended at once, and then
"filtered" so Postfix is okay with it.

Users can now override the default Postfix configuration and do not have
to worry about providing the config line-by-line.

The documentation was adjusted accordingly. It emphasises how the
configuration is applied to enable users to understand what they can do
and how they can achieve that.

Note: The new way of adjusting Postfix _should in theory_ be backwards
compatible. (Nothing is ever certain with Postfix...)